requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
connectorx==0.3.3
greenlet==3.1.0
idna==3.8
numpy==2.1.1
pandas==2.2.2
polars==0.19.8
psycopg==3.2.1
psycopg2==2.9.9
py4j==0.10.9.7
pyarrow==17.0.0
pyspark==3.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
